Moved the process runner and settings into the Common assembly. Made the required review changes on the comments.
