+++ src/devices/Camera/samples/Camera.Samples/Program.cs
@@ -1,12 +1,22 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Iot.Device.Camera;
 using Iot.Device.Camera.Settings;
+using Iot.Device.Common;
 
 namespace Camera.Samples;
 
 internal class Program
 {
+    private const string CmdList = "list";
+    private const string CmdStillLegacy = "still-legacy";
+    private const string CmdVideoLegacy = "video-legacy";
+    private const string CmdLapseLegacy = "lapse-legacy";
+    private const string CmdStillLibcamera = "still-libcamera";
+    private const string CmdVideoLibcamera = "video-libcamera";
+    private const string CmdLapseLibcamera = "lapse-libcamera";
+
     private static async Task<int> Main(string[] args)
     {
         if (args.Length == 0 || args.Length > 1)
@@ -17,13 +27,13 @@ internal class Program
         var arg = args[0];
         ProcessSettings? processSettings = arg switch
         {
-            "list" => ProcessSettingsFactory.CreateForLibcamerastillAndStderr(),
-            "still-legacy" => ProcessSettingsFactory.CreateForRaspistill(),
-            "video-legacy" => ProcessSettingsFactory.CreateForRaspivid(),
-            "lapse-legacy" => ProcessSettingsFactory.CreateForRaspistill(),
-            "still-libcamera" => ProcessSettingsFactory.CreateForLibcamerastill(),
-            "video-libcamera" => ProcessSettingsFactory.CreateForLibcameravid(),
-            "lapse-libcamera" => ProcessSettingsFactory.CreateForLibcamerastill(),
+            CmdList => ProcessSettingsFactory.CreateForLibcamerastillAndStderr(),
+            CmdStillLegacy => ProcessSettingsFactory.CreateForRaspistill(),
+            CmdVideoLegacy => ProcessSettingsFactory.CreateForRaspivid(),
+            CmdLapseLegacy => ProcessSettingsFactory.CreateForRaspistill(),
+            CmdStillLibcamera => ProcessSettingsFactory.CreateForLibcamerastill(),
+            CmdVideoLibcamera => ProcessSettingsFactory.CreateForLibcameravid(),
+            CmdLapseLibcamera => ProcessSettingsFactory.CreateForLibcamerastill(),
             _ => null,
         };
 
@@ -33,7 +43,7 @@ internal class Program
         }
 
         var capture = new Capture(processSettings);
-        if (arg == "list")
+        if (arg == CmdList)
         {
             var cams = await capture.List();
             Console.WriteLine("List of available cameras:");
@@ -45,7 +55,7 @@ internal class Program
             return 0;
         }
 
-        if (arg == "still-legacy" || arg == "still-libcamera")
+        if (arg == CmdStillLegacy || arg == CmdStillLibcamera)
         {
             var filename = await capture.CaptureStill();
             Console.WriteLine($"Captured the picture: {filename}");
@@ -53,7 +63,7 @@ internal class Program
             return 0;
         }
 
-        if (arg == "video-legacy" || arg == "video-libcamera")
+        if (arg == CmdVideoLegacy || arg == CmdVideoLibcamera)
         {
             var filename = await capture.CaptureVideo();
             Console.WriteLine($"Captured the video: {filename}");
@@ -61,7 +71,7 @@ internal class Program
             return 0;
         }
 
-        if (arg == "lapse-legacy" || arg == "lapse-libcamera")
+        if (arg == CmdLapseLegacy || arg == CmdLapseLibcamera)
         {
             await capture.CaptureTimelapse();
             Console.WriteLine($"The time-lapse images have been saved to disk");
@@ -75,13 +85,13 @@ internal class Program
     private static int Usage()
     {
         Console.WriteLine($"Camera.Samples supports one the following arguments:");
-        Console.WriteLine($"list             print the cameras available on (libcamera stack only)");
-        Console.WriteLine($"still-legacy     capture a still using raspistill");
-        Console.WriteLine($"video-legacy     capture a 10s video using raspivid");
-        Console.WriteLine($"lapse-legacy     capture a time lapse (5 images for 10s) using raspistill");
-        Console.WriteLine($"still-libcamera  capture a still using libcamera-still");
-        Console.WriteLine($"video-libcamera  capture a 10s video using libcamera-vid");
-        Console.WriteLine($"lapse-libcamera  capture a time lapse (5 images for 10s) using libcamera-still");
+        Console.WriteLine($"{CmdList}             print the cameras available on (libcamera stack only)");
+        Console.WriteLine($"{CmdStillLegacy}     capture a still using raspistill");
+        Console.WriteLine($"{CmdVideoLegacy}     capture a 10s video using raspivid");
+        Console.WriteLine($"{CmdLapseLegacy}     capture a time lapse (5 images for 10s) using raspistill");
+        Console.WriteLine($"{CmdStillLibcamera}  capture a still using libcamera-still");
+        Console.WriteLine($"{CmdVideoLibcamera}  capture a 10s video using libcamera-vid");
+        Console.WriteLine($"{CmdLapseLibcamera}  capture a time lapse (5 images for 10s) using libcamera-still");
         return -1;
     }
 }
