+++ src/devices/Common/Common.csproj
@@ -21,6 +21,7 @@
     <Compile Include="System/Device/Spi/*.cs" />
     <Compile Include="Interop/**/*.cs" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="$(MicrosoftExtensionsLoggingPackageVersion)" />
+    <PackageReference Include="System.IO.Pipelines" Version="$(SystemIOPipelinesVersion)" />
     <!-- The following is supported on windows only. We need it to take screenshots under windows -->
     <PackageReference Include="System.Drawing.Common" Version="$(SystemDrawingCommonPackageVersion)" />
   </ItemGroup>
